requirements.txt
openai
python-pptx
matplotlib
pandas
python-dotenv
gdown